Whenever searching Home i sno longer higlighted in sidebar.
