+++ src/components/SearchKeyWords.js
@@ -1,32 +1,22 @@
 import React from "react";
 import * as s from "./SearchKeyWords.cs";
+import SingleSearchKeyWord from "./SingleSearchKeyWord";
 
-function SearchKeyWords({ associatedKeywords }) {
+export default function SearchKeyWords({ associatedKeywords }) {
   let allKeywordsButLast = Object.entries(associatedKeywords).slice(0, -1);
   let lastKeyWord = Object.entries(associatedKeywords).pop();
 
   return (
     <s.ContainerKeywords>
       {allKeywordsButLast.map((associatedKeywords) => (
-        <Keywords
+        <SingleSearchKeyWord
           key={associatedKeywords[0]}
           text={associatedKeywords[1]}
           addSeperator={true}
         />
       ))}
 
-      <Keywords text={lastKeyWord[1]} />
+      <SingleSearchKeyWord text={lastKeyWord[1]} />
     </s.ContainerKeywords>
   );
 }
-
-function Keywords({ text, addSeperator }) {
-  return (
-    <s.ContainerKeywords>
-      <s.KeyWordText>{text}</s.KeyWordText>
-      {addSeperator && <hr style={{ color: "black" }}></hr>}
-    </s.ContainerKeywords>
-  );
-}
-
-export { SearchKeyWords, Keywords };
